requirements.txt
python-telegram-bot==20.8
pymongo==4.6.2
python-dotenv==1.0.1 